+++ saleor/graphql/attribute/filters.py
@@ -2,23 +2,34 @@ import django_filters
 import graphene
 from django.db.models import Q
 
+from ...attribute import AttributeInputType
 from ...attribute.models import Attribute, AttributeValue
 from ...permission.utils import has_one_of_permissions
 from ...product import models
 from ...product.models import ALL_PRODUCTS_PERMISSIONS
-from ..attribute.enums import AttributeTypeEnum
 from ..channel.filters import get_channel_slug_from_filter_data
+from ..core.descriptions import ADDED_IN_311, PREVIEW_FEATURE
+from ..core.enums import MeasurementUnitsEnum
 from ..core.filters import (
     EnumFilter,
     GlobalIDFilter,
     GlobalIDMultipleChoiceFilter,
     ListObjectTypeFilter,
     MetadataFilterBase,
+    OperationObjectTypeFilter,
     filter_slug_list,
 )
-from ..core.types import ChannelFilterInputObjectType, FilterInputObjectType
+from ..core.types import (
+    ChannelFilterInputObjectType,
+    FilterInputObjectType,
+    NonNullList,
+    StringFilterInput,
+)
+from ..core.types.filter_input import FilterInputDescriptions, WhereInputObjectType
 from ..core.utils import from_global_id_or_error
 from ..utils import get_user_or_app_from_context
+from ..utils.filters import filter_by_id, filter_by_string_field
+from .enums import AttributeEntityTypeEnum, AttributeInputTypeEnum, AttributeTypeEnum
 
 
 def filter_attributes_by_product_types(qs, field, value, requestor, channel_slug):
@@ -55,7 +66,13 @@ def filter_attributes_by_product_types(qs, field, value, requestor, channel_slug
     )
 
 
-def filter_attribute_type(qs, _, value):
+def filter_attribute_search(qs, _, value):
+    if not value:
+        return qs
+    return qs.filter(Q(slug__ilike=value) | Q(name__ilike=value))
+
+
+def filter_by_attribute_type(qs, _, value):
     if not value:
         return qs
     return qs.filter(type=value)
@@ -79,9 +96,9 @@ class AttributeValueFilter(django_filters.FilterSet):
 
 
 class AttributeFilter(MetadataFilterBase):
-    search = django_filters.CharFilter(method="filter_attribute_search")
+    search = django_filters.CharFilter(method=filter_attribute_search)
     ids = GlobalIDMultipleChoiceFilter(field_name="id")
-    type = EnumFilter(input_class=AttributeTypeEnum, method=filter_attribute_type)
+    type = EnumFilter(input_class=AttributeTypeEnum, method=filter_by_attribute_type)
 
     in_collection = GlobalIDFilter(method="filter_in_collection")
     in_category = GlobalIDFilter(method="filter_in_category")
@@ -97,9 +114,6 @@ class AttributeFilter(MetadataFilterBase):
             "filterable_in_dashboard",
             "available_in_grid",
         ]
-
-    def filter_attribute_search(self, qs, _, value):
-        return qs.filter(Q(slug__ilike=value) | Q(name__ilike=value))
 
     def filter_in_collection(self, qs, name, value):
         requestor = get_user_or_app_from_context(self.request)
@@ -124,3 +138,125 @@ class AttributeFilterInput(ChannelFilterInputObjectType):
 class AttributeValueFilterInput(FilterInputObjectType):
     class Meta:
         filterset_class = AttributeValueFilter
+
+
+class AttributeInputTypeEnumFilterInput(graphene.InputObjectType):
+    eq = AttributeInputTypeEnum(description=FilterInputDescriptions.EQ, required=False)
+    one_of = NonNullList(
+        AttributeInputTypeEnum,
+        description=FilterInputDescriptions.ONE_OF,
+        required=False,
+    )
+
+
+class AttributeEntityTypeEnumFilterInput(graphene.InputObjectType):
+    eq = AttributeEntityTypeEnum(description=FilterInputDescriptions.EQ, required=False)
+    one_of = NonNullList(
+        AttributeEntityTypeEnum,
+        description=FilterInputDescriptions.ONE_OF,
+        required=False,
+    )
+
+
+class AttributeTypeEnumFilterInput(graphene.InputObjectType):
+    eq = AttributeTypeEnum(description=FilterInputDescriptions.EQ, required=False)
+    one_of = NonNullList(
+        AttributeTypeEnum,
+        description=FilterInputDescriptions.ONE_OF,
+        required=False,
+    )
+
+
+class MeasurementUnitsEnumFilterInput(graphene.InputObjectType):
+    eq = MeasurementUnitsEnum(description=FilterInputDescriptions.EQ, required=False)
+    one_of = NonNullList(
+        MeasurementUnitsEnum,
+        description=FilterInputDescriptions.ONE_OF,
+        required=False,
+    )
+
+
+def filter_attribute_name(qs, _, value):
+    return filter_by_string_field(qs, "name", value)
+
+
+def filter_attribute_slug(qs, _, value):
+    return filter_by_string_field(qs, "slug", value)
+
+
+def filter_with_choices(qs, _, value):
+    lookup = Q(input_type__in=AttributeInputType.TYPES_WITH_CHOICES)
+    if value is True:
+        return qs.filter(lookup)
+    elif value is False:
+        return qs.exclude(lookup)
+    return qs
+
+
+def filter_attribute_input_type(qs, _, value):
+    return filter_by_string_field(qs, "input_type", value)
+
+
+def filter_attribute_entity_type(qs, _, value):
+    return filter_by_string_field(qs, "entity_type", value)
+
+
+def filter_attribute_type(qs, _, value):
+    return filter_by_string_field(qs, "type", value)
+
+
+def filter_attribute_unit(qs, _, value):
+    return filter_by_string_field(qs, "unit", value)
+
+
+class AttributeWhere(MetadataFilterBase):
+    ids = GlobalIDMultipleChoiceFilter(method=filter_by_id("Attribute"))
+    name = OperationObjectTypeFilter(
+        input_class=StringFilterInput, method=filter_attribute_name
+    )
+    slug = OperationObjectTypeFilter(
+        input_class=StringFilterInput, method=filter_attribute_slug
+    )
+    with_choices = django_filters.BooleanFilter(method=filter_with_choices)
+    input_type = OperationObjectTypeFilter(
+        AttributeInputTypeEnumFilterInput, method=filter_attribute_input_type
+    )
+    entity_type = OperationObjectTypeFilter(
+        AttributeEntityTypeEnumFilterInput, method=filter_attribute_entity_type
+    )
+    type = OperationObjectTypeFilter(
+        AttributeTypeEnumFilterInput, method=filter_attribute_type
+    )
+    unit = OperationObjectTypeFilter(
+        MeasurementUnitsEnumFilterInput, method=filter_attribute_unit
+    )
+    in_collection = GlobalIDFilter(method="filter_in_collection")
+    in_category = GlobalIDFilter(method="filter_in_category")
+
+    class Meta:
+        model = Attribute
+        fields = [
+            "value_required",
+            "visible_in_storefront",
+            "filterable_in_dashboard",
+        ]
+
+    def filter_in_collection(self, qs, name, value):
+        requestor = get_user_or_app_from_context(self.request)
+        channel_slug = get_channel_slug_from_filter_data(self.data)
+        return filter_attributes_by_product_types(
+            qs, name, value, requestor, channel_slug
+        )
+
+    def filter_in_category(self, qs, name, value):
+        requestor = get_user_or_app_from_context(self.request)
+        channel_slug = get_channel_slug_from_filter_data(self.data)
+        return filter_attributes_by_product_types(
+            qs, name, value, requestor, channel_slug
+        )
+
+
+class AttributeWhereInput(WhereInputObjectType):
+    class Meta:
+        filterset_class = AttributeWhere
+        description = "Where filtering options." + ADDED_IN_311 + PREVIEW_FEATURE