+++ saleor/graphql/attribute/types.py
@@ -198,6 +198,7 @@ class Attribute(ModelObjectType[models.Attribute]):
             f"{ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES.name}."
         ),
         required=True,
+        deprecation_reason=DEPRECATED_IN_3X_FIELD,
     )
     filterable_in_dashboard = graphene.Boolean(
         description=(
@@ -218,6 +219,7 @@ class Attribute(ModelObjectType[models.Attribute]):
             f"{ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES.name}."
         ),
         required=True,
+        deprecation_reason=DEPRECATED_IN_3X_FIELD,
     )
     storefront_search_position = graphene.Int(
         description=(
@@ -228,6 +230,7 @@ class Attribute(ModelObjectType[models.Attribute]):
             f"{ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES.name}."
         ),
         required=True,
+        deprecation_reason=DEPRECATED_IN_3X_FIELD,
     )
     translation = TranslationField(AttributeTranslation, type_name="attribute")
     with_choices = graphene.Boolean(