+++ saleor/graphql/page/tests/queries/test_page_type.py
@@ -7,7 +7,9 @@ from ....tests.utils import (
 )
 
 PAGE_TYPE_QUERY = """
-    query PageType($id: ID!, $filters: AttributeFilterInput) {
+    query PageType(
+        $id: ID!, $filters: AttributeFilterInput, $where: AttributeWhereInput
+    ) {
         pageType(id: $id) {
             id
             name
@@ -16,7 +18,7 @@ PAGE_TYPE_QUERY = """
             attributes {
                 slug
             }
-            availableAttributes(first: 10, filter: $filters) {
+            availableAttributes(first: 10, filter: $filters, where: $where) {
                 edges {
                     node {
                         slug
@@ -152,6 +154,40 @@ def test_page_type_query_filter_unassigned_attributes(
     variables = {
         "id": graphene.Node.to_global_id("PageType", page_type.pk),
         "filters": {"search": expected_attribute.name},
+    }
+
+    # when
+    response = staff_api_client.post_graphql(PAGE_TYPE_QUERY, variables)
+
+    # then
+    content = get_graphql_content(response)
+    page_type_data = content["data"]["pageType"]
+
+    assert page_type_data["slug"] == page_type.slug
+    assert {attr["slug"] for attr in page_type_data["attributes"]} == {
+        attr.slug for attr in page_type.page_attributes.all()
+    }
+    available_attributes = page_type_data["availableAttributes"]["edges"]
+    assert len(available_attributes) == 1
+    assert available_attributes[0]["node"]["slug"] == expected_attribute.slug
+
+
+def test_page_type_query_where_filter_unassigned_attributes(
+    staff_api_client,
+    page_type,
+    permission_manage_pages,
+    page_type_attribute_list,
+    color_attribute,
+):
+    # given
+    staff_user = staff_api_client.user
+    staff_user.user_permissions.add(permission_manage_pages)
+
+    expected_attribute = page_type_attribute_list[0]
+
+    variables = {
+        "id": graphene.Node.to_global_id("PageType", page_type.pk),
+        "where": {"name": {"eq": expected_attribute.name}},
     }
 
     # when